backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pdfplumber==0.11.0
groq==0.9.0
pydantic==2.7.1
python-multipart==0.0.9
aiofiles==23.2.1
python-dotenv==1.0.1
tqdm==4.66.4
httpx==0.27.0
# RAG stack
sentence-transformers==3.0.0
chromadb==0.5.3
torch==2.3.0
PyMuPDF
pytesseract
Pillow
rank_bm25
scikit-learn
